% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_palette.R
\name{cms_palette}
\alias{cms_palette}
\title{cms_palette
generates qualitative colour palette from RColorBrewer}
\usage{
cms_palette(n = 1, preview = FALSE, full = FALSE)
}
\arguments{
\item{n}{numeric; default 1. number of colours; cannot exceed 335}

\item{preview}{logic; default FALSE; when set, gives preview of colour}

\item{full}{logic; default FALSE; when set, returns all 355 colours}
}
\value{
vector of HEX colours, if full set to TRUE, gives all 335 colours, else gives n numbers
}
\description{
cms_palette
generates qualitative colour palette from RColorBrewer
}
