% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_rarefy.R
\name{cms_rarefy}
\alias{cms_rarefy}
\title{cms_rarefy}
\usage{
cms_rarefy(x)
}
\arguments{
\item{x}{matrix with samples in row, feature in column}
}
\description{
rarefaction curve based on vegan. returns dataframe of rarefaction curve adapted from https://github.com/mahendra-mariadassou/phyloseq-extended/blob/master/R/graphical_methods.R and vegan
}
