% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_da_prop.R
\name{mod_da_prop_ui}
\alias{mod_da_prop_ui}
\alias{mod_da_prop_server}
\title{mod_da_prop_ui and mod_da_prop_server}
\usage{
mod_da_prop_ui(id)

mod_da_prop_server(input, output, session, param)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{param}{reactive values to communicate with outer module}
}
\description{
A shiny submodule of differential abundance analysis.
Calculates and visuzlaizes proportionality scores
}
\keyword{internal}
