% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_diffAbund.R
\name{mod_diffAbund_ui}
\alias{mod_diffAbund_ui}
\alias{mod_diffAbund_server}
\title{mod_diffAbund_ui and mod_diffAbund_server}
\usage{
mod_diffAbund_ui(id)

mod_diffAbund_server(input, output, session, improxy)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{improxy}{import reactive values across modules}
}
\description{
A shiny Module.
}
\keyword{internal}
