% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explore.R, R/mod_overview.R
\name{mod_explore_ui}
\alias{mod_explore_ui}
\alias{mod_explore_server}
\alias{mod_overview_server}
\title{mod_explore_ui and mod_explore_server}
\usage{
mod_explore_ui(id)

mod_explore_server(input, output, session, improxy)

mod_overview_server(input, output, session, improxy)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
Overview of dataset using exploratory analysis
}
\keyword{internal}
