% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_import.R
\name{mod_import_ui}
\alias{mod_import_ui}
\alias{mod_import_server}
\title{mod_import_ui and mod_import_server}
\usage{
mod_import_ui(id)

mod_import_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
Import dataset and show basic overview
}
\keyword{internal}
