% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_intro.R
\name{mod_intro_ui}
\alias{mod_intro_ui}
\alias{mod_intro_server}
\title{mod_intro_ui and mod_intro_server}
\usage{
mod_intro_ui(id)

mod_intro_server(input, output, session, parent_session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
