% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ov_hmap.R
\name{mod_ov_hmap_ui}
\alias{mod_ov_hmap_ui}
\alias{mod_ov_hmap_server}
\title{mod_ov_hmap_ui and mod_ov_hmap_server}
\usage{
mod_ov_hmap_ui(id)

mod_ov_hmap_server(input, output, session, param)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
