% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ov_pcoa.R
\name{mod_ov_pcoa_ui}
\alias{mod_ov_pcoa_ui}
\alias{mod_ov_pcoa_server}
\title{mod_ov_pcoa_ui and mod_ov_pcoa_server}
\usage{
mod_ov_pcoa_ui(id)

mod_ov_pcoa_server(input, output, session, bridge)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
