% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_qc.R
\name{mod_qc_ui}
\alias{mod_qc_ui}
\alias{mod_qc_server}
\title{mod_qc_ui and mod_qc_server}
\usage{
mod_qc_ui(id)

mod_qc_server(input, output, session, improxy)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
Provides interactive version of QC report returned from dada2 
analysis via CGAT-core pipeline
}
\keyword{internal}
