% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_setup.R
\name{mod_setup_ui}
\alias{mod_setup_ui}
\alias{mod_setup_server}
\title{mod_setup_ui and mod_setup_server}
\usage{
mod_setup_ui(id)

mod_setup_server(input, output, session, improxy)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
Prepare data for analysis
}
\keyword{internal}
