% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_dendro.R
\name{plot_ggdendro}
\alias{plot_ggdendro}
\title{plot_ggdendro}
\usage{
plot_ggdendro(
  hcdata,
  metadata,
  direction = c("lr", "rl", "tb", "bt"),
  fan = FALSE,
  branch.size = 1,
  label.size = 3,
  label.category = NULL,
  nudge.label = 0.01,
  expand.y = 0.5,
  category = NULL,
  id = "sampleID"
)
}
\arguments{
\item{hcdata}{dendrogram data}

\item{metadata}{dataframe containing metadata used to annotate dendrogram}

\item{direction}{direction of level tob-bottom ('tb'), bottom-top ('bt'),
left-right ('lr'), right-left ('rl')}

\item{fan}{fan the dendrogram}

\item{branch.size}{branch size}

\item{label.size}{label size}

\item{label.category}{colour label by}

\item{nudge.label}{nudge label spacing}

\item{expand.y}{expand y axis limit to this value}

\item{category}{colour dendrogram leaves by}

\item{id}{variable to label dendrogram by}
}
\description{
use ggplot to make customized dendrogram.
following https://atrebas.github.io/post/2019-06-08-lightweight-dendrograms/
}
