% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_helper.R
\name{withBusyIndicatorServer}
\alias{withBusyIndicatorServer}
\title{withBusyIndicatorServer}
\usage{
withBusyIndicatorServer(buttonId, mod_name, expr)
}
\arguments{
\item{mod_name}{name of shiny module where \code{withBusyIndicatorServer} is called}

\item{expr}{code executed when button clicked}

\item{buttonID}{button id}
}
\description{
Call this function from the server with the button id that is clicked and the
expression to run when the button is clicked.
Corresponds with \code{withBustyIndicatorUI} function.
Copied and modified from \href{https://github.com/daattali/advanced-shiny/blob/master/busy-indicator/helpers.R}
}
