% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_helper.R
\name{withBusyIndicatorUI}
\alias{withBusyIndicatorUI}
\title{withBusyIndicatorUI}
\usage{
withBusyIndicatorUI(button)
}
\arguments{
\item{button}{Action button for indicator to act on}
}
\description{
UI function to show busy/loading/error indicator after button is pressed.
Corresponds with \code{withBusyIndicatorServer} function.
copied from \href{https://github.com/daattali/advanced-shiny/blob/master/busy-indicator/helpers.R}
}
