% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_rarefy.R
\name{cms_rarefy}
\alias{cms_rarefy}
\title{cms_rarefy}
\usage{
cms_rarefy(df)
}
\arguments{
\item{x}{matrix with samples in columns, feature in rows
import foreach
import vegan}
}
\description{
from MetaSequencing by microbialman
}
