% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create_db}
\alias{create_db}
\title{create_db
generates qualitative colour palette from RColorBrewer}
\usage{
create_db(counts, taxonomy, overwrite = FALSE, outdir)
}
\arguments{
\item{counts}{dataframe of counts with samples in columns, features in rows. 
First column should contain feature identities under the heading of 
\code{featureID}. Feature IDs must be unique.}

\item{taxonomy}{dataframe of taxonomy classifications corresponding to the 
counts table. Must contain the columns \code{featureID}, \code{sequence}, 
\code{Kingdom}, \code{Phylum}, \code{Class}, \code{Order}, \code{Family},
\code{Genus}, \code{Species}, \code{Taxon}. \code{Taxon} contains 
taxonomic lassifications at all taxonomic levels, seperated by a semicolon. 
Unclassified taxa should be represented with \code{NA}.}

\item{overwrite}{logic. default FALSE. Allow database tables to overwrite
existing tables of the same name.}

\item{outdir}{output directory}
}
\value{
sqlite database with merged_abundance_id (counts) and
  merged_taxonomy (taxonomy) in a sqlite database file named 'csvdb'
}
\description{
create_db
generates qualitative colour palette from RColorBrewer
}
