% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_dendro.R
\name{dendro_data_k}
\alias{dendro_data_k}
\title{dendro_data_k}
\usage{
dendro_data_k(hc, k)
}
\description{
Extract dendrogram data for plot customized dendrogram. 
following \href{https://atrebas.github.io/post/2019-06-08-lightweight-dendrograms/}
}
