% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_helper.R
\name{withBusyIndicatorServer}
\alias{withBusyIndicatorServer}
\title{withBusyIndicatorServer}
\usage{
withBusyIndicatorServer(buttonId, expr)
}
\description{
Call this function from the server with the button id that is clicked and the
expression to run when the button is clicked.
Corresponds with \code{withBustyIndicatorUI} function.
Copied from \link{https://github.com/daattali/advanced-shiny/blob/master/busy-indicator/helpers.R}
}
